[actions] Move Docker to main package since now it will be used by both source and IR
